+++ src/lib/_interfaces.ts
@@ -1,3 +1,10 @@
+/**
+ * Represents a lazy value of type `T`.
+ * You can retrieve it with `const val = typeof lazy === "function" ? lazy() : lazy;`.
+ * This library guarantees that it will be only called once but you should still ensure that it is idempotent.
+ */
+export type Lazy<T> = T | (() => T);
+
 export interface Type<T> {
   name: string;
   testError (val: T): Error | undefined;