+++ src/test/circular-types/fs-tree/fs-node-type.ts
@@ -1,0 +1,12 @@
+import {CaseStyle} from "../../../lib/_helpers/rename";
+import {SimpleEnumType} from "../../../lib/simple-enum";
+
+export enum FsNodeType {
+  File,
+  Directory,
+}
+
+export const $FsNodeType: SimpleEnumType<FsNodeType> = new SimpleEnumType<FsNodeType>({
+  enum: FsNodeType,
+  rename: CaseStyle.KebabCase,
+});