+++ src/test/circular-types/fs-tree/fs-node.ts
@@ -1,0 +1,15 @@
+import {TaggedUnionType} from "../../../lib/tagged-union";
+import {$Directory, Directory} from "./directory";
+import {$File, File} from "./file";
+
+export type FsNode =
+  Directory
+  | File;
+
+export const $FsNode: TaggedUnionType<FsNode> = new TaggedUnionType<FsNode>({
+  variants: [
+    $Directory,
+    $File,
+  ],
+  tag: "type",
+});